Refactor and Update Codebase
- Updated C++ compiler flags in Makefile to include -Wall, -Wextra, and -Wpedantic for better code quality checks.
- Changed struct declarations to class in parser.y and parser.tab.h for TokenAttribute and ASTNode to improve encapsulation.
- Enhanced user feedback in main function by replacing Unicode characters with emojis for better readability.
- Commented out unused ADDRESS_VAR and isAddressSymbol function in header files to clean up the codebase.
- Modified input.c to prevent out-of-bounds access by changing the assignment of arr[7] to a valid variable declaration.
- Updated loop indices from int to size_t in various source files for better type safety and to avoid potential issues with large sizes.
- Removed unnecessary address space checks in code generation functions to simplify logic.
- Cleaned up unused variables and commented-out code in various files to improve maintainability.

+++ Stages/End-to-End/src/controlFlowGraph.cpp
@@ -140,7 +140,7 @@ int CFG::whichBlock(int index)
 {
     // Find the just smaller leader than index
     int foundLeader = -1;
-    for (int i = 0; i < leaders.size(); i++)
+    for (size_t i = 0; i < leaders.size(); i++)
     {
         if (leaders[i] <= index)
         {
